requirements.txt
altgraph==0.15
dis3==0.1.2
future==0.16.0
macholib==1.9
pefile==2017.11.5
PyInstaller==3.3.1
pyserial==2.7
pywin32==223
